cleaned directories
