+++ src/ESP_testing_1D/ESP_testing_1D.ino
@@ -1,29 +1,23 @@
-/*
-  PC to ESP module for communication using telnet
-*/
-
-//if using in a different folder remember to go to preference and change the folder default location
-
 #include <ESP8266WiFi.h>
 
-#define SendKey 0  //Button to send data Flash BTN on NodeMCU
+#define BAUD 115200 
 
 int port = 8888;  //Port number
 WiFiServer server(port);
+// Set random static IP address, gateway IP address, subnet
+IPAddress local_IP(192, 168, 1, 2); //ip to which pc should connect
+IPAddress gateway(192, 168, 1, 1); //randomly allocated
+IPAddress subnet(255, 255, 0, 0); //randomly allocated
 
 //Server connect to WiFi Network
 const char *ssid = "NETGEAR35";        //Enter your wifi SSID
-const char *password = "change_me";  //Enter your wifi Password
+const char *password = "amberlab";  //Enter your wifi Password
 
 int count=0;
 
-//=======================================================================
-//                    Power on setup
-//=======================================================================
 void setup() 
 {
-  Serial.begin(115200);
-  pinMode(SendKey,INPUT_PULLUP);  //Btn to send data
+  Serial.begin(BAUD); //setup baud rate must be the same as for the pc receiver
   Serial.println();
 
   WiFi.mode(WIFI_STA);
@@ -42,19 +36,23 @@ void setup()
   Serial.println(ssid);
 
   Serial.print("IP address: ");
-  Serial.println(WiFi.localIP());  //e.g. 192.168.1.3
+  Serial.println(WiFi.localIP()); //shouldnt we spacify here address again?
   server.begin();
   Serial.print("Open Telnet and connect to IP:");
   Serial.print(WiFi.localIP());
   Serial.print(" on port ");
   Serial.println(port);
 }
-//=======================================================================
-//                    Loop
-//=======================================================================
 
 void loop() 
 {
+  //=======================================================
+  //                 RX/TX RECEPTION
+  //=======================================================
+  
+  //=======================================================
+  //                 WIFI SENDING
+  //=======================================================
   WiFiClient client = server.available();
   
   if (client) {
@@ -78,4 +76,3 @@ void loop()
     Serial.println("Client disconnected");    
   }
 }
-//=======================================================================
