+++ src/PC_testing_fun.h
@@ -1,0 +1,6 @@
+#ifndef __PC_testing_fun_h__
+#define  __PC_testing_fun_h__ 
+
+void func(int sockfd);
+
+#endif
