+++ src/makefile
@@ -1,0 +1,2 @@
+pc_testing_make: PC_testing_1D.c PC_testing_fun.c
+	gcc -o pc_testing_make PC_testing_1D.c PC_testing_fun.c -I.
